requirements.txt
# Data Quality (Baru ditambahkan)
great_expectations==1.15.1

# Orchestration (Baru ditambahkan)
apache-airflow==2.10.4

# Streaming
kafka-python==2.3.0

# Data processing
pandas==3.0.1
pyarrow==23.0.1
pyspark==4.1.1

# Transform
dbt-core==1.11.7
dbt-duckdb==1.10.1
duckdb==1.5.0

# Serving
fastapi==0.135.1
uvicorn==0.42.0

# Utilities
faker==40.11.0
